[cr134] Whitelist native mojom types for migration
Mojom use of `[native]` is being abandoned, and it is supposed to be
migrated away from. In the meantime, types using this attribute are
supposed to be added to an allow list. This change adds the two cases in
brave to the patch for the python file with the list.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/2beb4744e13743f13dcdabf2cb0e09899886fe60

commit 2beb4744e13743f13dcdabf2cb0e09899886fe60
Author: Daniel Cheng <dcheng@chromium.org>
Date:   Mon Feb 3 08:10:45 2025 -0800

    Add allowlist for using [Native]

    This was primarily an escape hatch to faciliate the migration of legacy
    IPC to mojo IPC. This is no longer really needed, an no new instances
    should be added.

    Bug: 393179188